feat(visits): add upcoming, successful, and cancelled visits endpoints; update visit staff route; remove redundant staff update from client routes
